Implement the read skew scenario
The original two papers seem to only consider the situation of only one
session. For now assume the first step of every session take a snapshot
of the initial state, and retake it at the beginning of rest of the steps
if the isolation level is read committed. Wonder if it is theoretically
sound.